refactored Bubble class as own model, updated logic accordingly, increased bubble sizing
