+++ src/App.css
@@ -64,8 +64,8 @@ div {
 
 canvas {
   background: #56789a;
-  width: 100%;
-  height: 100%;
+  width: 85vw;
+  height: 85vh;
   scroll-snap-align: start;
 }
